scripts to add zong and telenor bundles in database get bundles api conversion from string to jsonarraynode and string to jsonnode

+++ src/main/java/com/zingpay/repository/BundleRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
  */
 
 @Repository
-public interface BundleRepository extends BaseRepository<Bundle, Long> {
+public interface BundleRepository extends BaseRepository<Bundle, Integer> {
     List<Bundle> findAllByNetwork(String network);
 }
